Issue #68: moved test function to config folder [ci skip]
